refactor: consolidate utility modules under util/
Move path, limits, ptr_arc, and executor into a new util/ submodule
to reduce clutter at the crate root. No semantic changes.
