+++ crates/pixi_command_dispatcher/src/util/limits.rs
