+++ crates/pixi_command_dispatcher/src/util/path.rs
