+++ crates/pixi_command_dispatcher/src/util/ptr_arc.rs
